src/deployment/web_app/requirements.txt
streamlit
azure-identity
azure-storage-blob
azure-keyvault
htbuilder
streamlit-authenticator>=0.2
pyyaml
numpy==1.23.5